+++ lib/src/main/java/it/unicam/quasylab/jspear/controller/ProbabilisticInterleavingController.java
@@ -36,7 +36,7 @@ import java.util.List;
  */
 public class ProbabilisticInterleavingController implements Controller {
 
-    private final int p;
+    private final double p;
     private final Controller leftController;
     private final Controller rightController;
 
@@ -47,7 +47,7 @@ public class ProbabilisticInterleavingController implements Controller {
      * @param leftController a controller.
      * @param rightController a controller.
      */
-    public ProbabilisticInterleavingController(int p, Controller leftController, Controller rightController) {
+    public ProbabilisticInterleavingController(double p, Controller leftController, Controller rightController) {
         if (p<0 || p>1) {
             throw new IllegalArgumentException("Illegal probability weight!");
         }
